Fix(Cart): arrange the cart component to match the figma design
